Add refresh btn EventListener and re-organice code
